django_app/requirements.txt
Django>=3.0,<4.0
psycopg2-binary>=2.8
django-crispy-forms==1.9.2
numpy==1.19.2
joblib==0.16.0 
scikit-learn==0.22.2 
scipy==1.5.2 
sklearn==0.0 
threadpoolctl==2.1.0
djangorestframework==3.12.1
cffi==1.14.3
cryptography==3.1.1
django-rest-knox==4.1.0 
pycparser==2.20 
six==1.15.0